+++ noodles-refget/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * refget: Increase the visibility of the `sequence` module.
+
 ## 0.2.0 - 2023-12-14
 
 ### Changed